+++ test/core-e2e/src/plugins.spec.js
@@ -110,14 +110,10 @@ describe("plugins", () => {
           ).toEqual(1);
         });
 
-        it("should have passed core to init method", async () => {
-          expect(initSpy.getCall(0).args[0]).toEqual(core);
-        });
-
         it("should have config available when init is called", async () => {
-          expect(initSpy.getCall(0).args[1]).toEqual(fixturesFolder("web-tutorial"));
-          expect(initSpy.getCall(0).args[2]).toEqual(3100);
-          expect(initSpy.getCall(0).args[3]).toEqual(0);
+          expect(initSpy.getCall(0).args[0]).toEqual(fixturesFolder("web-tutorial"));
+          expect(initSpy.getCall(0).args[1]).toEqual(3100);
+          expect(initSpy.getCall(0).args[2]).toEqual(0);
         });
 
         it("should have executed start method", async () => {
@@ -128,10 +124,6 @@ describe("plugins", () => {
           expect(
             filterLogs(core.logs, "[plugins:test-plugin] Log from start method").length
           ).toEqual(1);
-        });
-
-        it("should have passed core to start method", async () => {
-          expect(startSpy.getCall(0).args[0]).toEqual(core);
         });
 
         it("should respond to custom routes", async () => {
@@ -233,29 +225,28 @@ describe("plugins", () => {
 
   testPlugin("created as an object", {
     id: "test-plugin",
-    register: ({ core: coreInstance, alerts, config, logger }) => {
+    register: ({ addRouter, alerts, config, logger }) => {
+      registerSpy();
       logger.info("Log from register method");
-      coreInstance.addRouter("/foo-path", customRouter);
+      addRouter("/foo-path", customRouter);
       alerts.set("test-register", "Warning registering plugin");
-      registerSpy(coreInstance);
       configSpy(config);
     },
-    init: ({ core: coreInstance, logger }) => {
+    init: ({ onChangeAlerts, onChangeMocks, config, logger }) => {
       logger.info("Log from init method");
       initSpy(
-        coreInstance,
-        coreInstance.config.namespace("files").option("path").value,
-        coreInstance.config.namespace("server").option("port").value,
-        coreInstance.config.namespace("mocks").option("delay").value
+        config.root.namespace("files").option("path").value,
+        config.root.namespace("server").option("port").value,
+        config.root.namespace("mocks").option("delay").value
       );
-      coreInstance.config.option("log").value = "silly";
-      coreInstance.onChangeAlerts(changeAlertsSpy);
-      coreInstance.onChangeMocks(mocksLoadedSpy);
+      config.root.option("log").value = "silly";
+      onChangeAlerts(changeAlertsSpy);
+      onChangeMocks(mocksLoadedSpy);
     },
-    start: ({ core: coreInstance, alerts, logger }) => {
+    start: ({ alerts, logger }) => {
       logger.info("Log from start method");
-      startSpy(coreInstance);
       alerts.set("test-start", "Warning starting plugin");
+      startSpy();
     },
     stop: ({ alerts, logger }) => {
       logger.info("Log from stop method");
@@ -286,30 +277,28 @@ describe("plugins", () => {
         return "test-plugin";
       }
 
-      constructor({ core: coreInstance, alerts, config, logger }) {
+      constructor({ addRouter, alerts, config, logger }) {
         logger.info("Log from register method");
-        this._core = coreInstance;
-        coreInstance.addRouter("/foo-path", customRouter);
+        addRouter("/foo-path", customRouter);
         alerts.set("test-register", "Warning registering plugin");
-        registerSpy(coreInstance);
+        registerSpy();
         configSpy(config);
       }
-      init({ core: coreInstance, logger }) {
+      init({ onChangeAlerts, onChangeMocks, logger, config }) {
         logger.info("Log from init method");
         initSpy(
-          coreInstance,
-          coreInstance.config.namespace("files").option("path").value,
-          coreInstance.config.namespace("server").option("port").value,
-          coreInstance.config.namespace("mocks").option("delay").value
+          config.root.namespace("files").option("path").value,
+          config.root.namespace("server").option("port").value,
+          config.root.namespace("mocks").option("delay").value
         );
-        coreInstance.config.option("log").value = "silly";
-        coreInstance.onChangeAlerts(changeAlertsSpy);
-        coreInstance.onChangeMocks(mocksLoadedSpy);
+        config.root.option("log").value = "silly";
+        onChangeAlerts(changeAlertsSpy);
+        onChangeMocks(mocksLoadedSpy);
       }
-      start({ core: coreInstance, alerts, logger }) {
+      start({ alerts, logger }) {
         logger.info("Log from start method");
         alerts.set("test-start", "Warning starting plugin");
-        startSpy(coreInstance);
+        startSpy();
       }
       stop({ alerts, logger }) {
         logger.info("Log from stop method");
@@ -325,29 +314,28 @@ describe("plugins", () => {
         return "test-plugin";
       }
 
-      register({ core: coreInstance, alerts, config, logger }) {
+      register({ addRouter, alerts, config, logger }) {
         logger.info("Log from register method");
-        coreInstance.addRouter("/foo-path", customRouter);
+        addRouter("/foo-path", customRouter);
         alerts.set("test-register", "Warning registering plugin");
-        registerSpy(coreInstance);
+        registerSpy();
         configSpy(config);
       }
-      init({ core: coreInstance, logger }) {
+      init({ logger, config, onChangeAlerts, onChangeMocks }) {
         logger.info("Log from init method");
         initSpy(
-          coreInstance,
-          coreInstance.config.namespace("files").option("path").value,
-          coreInstance.config.namespace("server").option("port").value,
-          coreInstance.config.namespace("mocks").option("delay").value
+          config.root.namespace("files").option("path").value,
+          config.root.namespace("server").option("port").value,
+          config.root.namespace("mocks").option("delay").value
         );
-        coreInstance.config.option("log").value = "silly";
-        coreInstance.onChangeAlerts(changeAlertsSpy);
-        coreInstance.onChangeMocks(mocksLoadedSpy);
+        config.root.option("log").value = "silly";
+        onChangeAlerts(changeAlertsSpy);
+        onChangeMocks(mocksLoadedSpy);
       }
-      start({ core: coreInstance, alerts, logger }) {
+      start({ alerts, logger }) {
         logger.info("Log from start method");
         alerts.set("test-start", "Warning starting plugin");
-        startSpy(coreInstance);
+        startSpy();
       }
       stop({ alerts, logger }) {
         logger.info("Log from stop method");
@@ -359,29 +347,28 @@ describe("plugins", () => {
   testPlugin(
     "created as a Class with register method and without static id",
     class Plugin {
-      register({ core: coreInstance, alerts, config, logger }) {
+      register({ addRouter, alerts, config, logger }) {
         logger.info("Log from register method");
-        coreInstance.addRouter("/foo-path", customRouter);
+        addRouter("/foo-path", customRouter);
         alerts.set("test-register", "Warning registering plugin");
-        registerSpy(coreInstance);
+        registerSpy();
         configSpy(config);
       }
-      init({ core: coreInstance, logger }) {
+      init({ logger, config, onChangeAlerts, onChangeMocks }) {
         logger.info("Log from init method");
         initSpy(
-          coreInstance,
-          coreInstance.config.namespace("files").option("path").value,
-          coreInstance.config.namespace("server").option("port").value,
-          coreInstance.config.namespace("mocks").option("delay").value
+          config.root.namespace("files").option("path").value,
+          config.root.namespace("server").option("port").value,
+          config.root.namespace("mocks").option("delay").value
         );
-        coreInstance.config.option("log").value = "silly";
-        coreInstance.onChangeAlerts(changeAlertsSpy);
-        coreInstance.onChangeMocks(mocksLoadedSpy);
+        config.root.option("log").value = "silly";
+        onChangeAlerts(changeAlertsSpy);
+        onChangeMocks(mocksLoadedSpy);
       }
-      start({ core: coreInstance, alerts, logger }) {
+      start({ alerts, logger }) {
         logger.info("Log from start method");
         alerts.set("test-start", "Warning starting plugin");
-        startSpy(coreInstance);
+        startSpy();
       }
       stop({ alerts, logger }) {
         logger.info("Log from stop method");
@@ -400,10 +387,10 @@ describe("plugins", () => {
         return "test-plugin";
       }
 
-      register({ core: coreInstance, alerts, config }) {
-        coreInstance.addRouter("/foo-path", customRouter);
+      register({ addRouter, alerts, config }) {
+        addRouter("/foo-path", customRouter);
         alerts.set("test-register", "Warning registering plugin");
-        registerSpy(coreInstance);
+        registerSpy();
         configSpy(config);
       }
       async init() {
@@ -416,9 +403,9 @@ describe("plugins", () => {
     }
   );
 
-  testPlugin("created as a function", ({ core: coreInstance }) => {
-    coreInstance.addRouter("/foo-path", customRouter);
-    registerSpy(coreInstance);
+  testPlugin("created as a function", ({ addRouter }) => {
+    addRouter("/foo-path", customRouter);
+    registerSpy();
     return {
       id: "test-plugin",
       register({ config, alerts, logger }) {
@@ -426,22 +413,21 @@ describe("plugins", () => {
         configSpy(config);
         alerts.set("test-register", "Warning registering plugin");
       },
-      init: ({ core: coreIns, logger }) => {
+      init: ({ config, logger, onChangeAlerts, onChangeMocks }) => {
         logger.info("Log from init method");
         initSpy(
-          coreIns,
-          coreIns.config.namespace("files").option("path").value,
-          coreIns.config.namespace("server").option("port").value,
-          coreIns.config.namespace("mocks").option("delay").value
+          config.root.namespace("files").option("path").value,
+          config.root.namespace("server").option("port").value,
+          config.root.namespace("mocks").option("delay").value
         );
-        coreIns.config.option("log").value = "silly";
-        coreIns.onChangeAlerts(changeAlertsSpy);
-        coreIns.onChangeMocks(mocksLoadedSpy);
+        config.root.option("log").value = "silly";
+        onChangeAlerts(changeAlertsSpy);
+        onChangeMocks(mocksLoadedSpy);
       },
       start: ({ alerts: alertsHere, logger }) => {
         logger.info("Log from start method");
         alertsHere.set("test-start", "Warning starting plugin");
-        startSpy(coreInstance);
+        startSpy();
       },
       stop: ({ alerts: alertsHere, logger }) => {
         logger.info("Log from stop method");
@@ -452,9 +438,9 @@ describe("plugins", () => {
 
   testAsyncPlugin(
     "created as a function",
-    ({ core: coreInstance }) => {
-      coreInstance.addRouter("/foo-path", customRouter);
-      registerSpy(coreInstance);
+    ({ addRouter }) => {
+      addRouter("/foo-path", customRouter);
+      registerSpy();
       return {
         register: ({ alerts }) => {
           alerts.set("test-register", "Warning registering plugin");
@@ -475,29 +461,28 @@ describe("plugins", () => {
   testPlugin("created as a function returning register property", () => {
     return {
       id: "test-plugin",
-      register: ({ core: coreInstance, alerts, config, logger }) => {
+      register: ({ addRouter, alerts, config, logger }) => {
         logger.info("Log from register method");
-        coreInstance.addRouter("/foo-path", customRouter);
+        addRouter("/foo-path", customRouter);
         alerts.set("test-register", "Warning registering plugin");
-        registerSpy(coreInstance);
+        registerSpy();
         configSpy(config);
       },
-      init: ({ core: coreInstance, logger }) => {
+      init: ({ logger, config, onChangeAlerts, onChangeMocks }) => {
         logger.info("Log from init method");
         initSpy(
-          coreInstance,
-          coreInstance.config.namespace("files").option("path").value,
-          coreInstance.config.namespace("server").option("port").value,
-          coreInstance.config.namespace("mocks").option("delay").value
+          config.root.namespace("files").option("path").value,
+          config.root.namespace("server").option("port").value,
+          config.root.namespace("mocks").option("delay").value
         );
-        coreInstance.config.option("log").value = "silly";
-        coreInstance.onChangeAlerts(changeAlertsSpy);
-        coreInstance.onChangeMocks(mocksLoadedSpy);
+        config.root.option("log").value = "silly";
+        onChangeAlerts(changeAlertsSpy);
+        onChangeMocks(mocksLoadedSpy);
       },
-      start: ({ core: coreInstance, alerts, logger }) => {
+      start: ({ alerts, logger }) => {
         logger.info("Log from start method");
         alerts.set("test-start", "Warning starting plugin");
-        startSpy(coreInstance);
+        startSpy();
       },
       stop: ({ alerts, logger }) => {
         logger.info("Log from stop method");