+++ packages/config/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
-## [Unreleased]
+## [1.1.0] - 2022-06-27
 
 ### Added
 - feat: Add `root` getter returning root config object to namespaces. Add it also to root config as an alias in order to keep the same interface.